+++ src/static/html/index.html
@@ -2,7 +2,7 @@
     <head>
         <title>Game Score API</title>
         <script type="text/javascript" src="/static/js/jquery.js"></script>
-        <script type="text/javascript" src="/static/js/index.js"></script>
+        <script type="module" src="/static/js/index.js"></script>
         <link rel="stylesheet" type="text/css" href="/static/css/bootstrap.min.css">
         <link rel="stylesheet" type="text/css" href="/static/css/app.css">
