+++ src/static/js/utils.js
@@ -1,0 +1,5 @@
+function renderError(response) {
+    console.error(response);
+}
+
+export { renderError }
